refactor: dividir ConfigurableForm y Form en archivos separados
- Se mueve la implementación de ConfigurableForm al archivo `configurable_form.rs`.
