+++ src/config/mod.rs
@@ -1,0 +1,1 @@
+pub mod configurable_form;
