+++ src/main.rs
@@ -1,5 +1,6 @@
 mod customs;
 mod form;
+mod config;
 use log::{info};
 
 use crate::form::Form;
